User Monitoring/assignment/requirements.txt
asgiref==3.2.10
cfe==0.0.15
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
gunicorn==20.0.4
Markdown==3.3.2
numpy==1.19.2
pandas==1.1.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
rest-pandas==1.1.0
six==1.15.0
sqlparse==0.4.1
